requirements.txt
yfinance==0.1.70
discord.py==1.7.3
async-timeout==3.0.1
